Moved the driver resource classes into a proper resources namespace
